Refactor: Move tests for new UI to viewModel test class
