+++ WordPress/src/test/java/org/wordpress/android/ui/sitecreation/domains/SiteCreationDomainsViewModelTest.kt
@@ -30,6 +30,8 @@ import org.wordpress.android.fluxc.store.SiteStore.SuggestDomainError
 import org.wordpress.android.ui.sitecreation.domains.SiteCreationDomainsViewModel.DomainModel
 import org.wordpress.android.ui.sitecreation.domains.SiteCreationDomainsViewModel.DomainsUiState
 import org.wordpress.android.ui.sitecreation.domains.SiteCreationDomainsViewModel.DomainsUiState.DomainsUiContentState
+import org.wordpress.android.ui.sitecreation.domains.SiteCreationDomainsViewModel.ListItemUiState.New
+import org.wordpress.android.ui.sitecreation.domains.SiteCreationDomainsViewModel.ListItemUiState.New.DomainUiState.Cost
 import org.wordpress.android.ui.sitecreation.domains.SiteCreationDomainsViewModel.ListItemUiState.Old
 import org.wordpress.android.ui.sitecreation.domains.SiteCreationDomainsViewModel.ListItemUiState.Old.DomainUiState.UnavailableDomain
 import org.wordpress.android.ui.sitecreation.misc.SiteCreationTracker
@@ -39,6 +41,7 @@ import org.wordpress.android.ui.sitecreation.usecases.FetchDomainsUseCase
 import org.wordpress.android.ui.utils.UiString.UiStringRes
 import org.wordpress.android.util.NetworkUtilsWrapper
 import org.wordpress.android.util.config.SiteCreationDomainPurchasingFeatureConfig
+import kotlin.test.assertIs
 
 private const val MULTI_RESULT_DOMAIN_FETCH_RESULT_SIZE = 20
 private val MULTI_RESULT_DOMAIN_FETCH_QUERY = "multi_result_query" to MULTI_RESULT_DOMAIN_FETCH_RESULT_SIZE
@@ -322,6 +325,86 @@ class SiteCreationDomainsViewModelTest : BaseUnitTest() {
             eq(MULTI_RESULT_DOMAIN_FETCH_QUERY.second),
         )
     }
+
+    // region New UI
+
+    private fun <T> testWithSuccessResultNewUi(
+        queryToSize: Pair<String, Int> = MULTI_RESULT_DOMAIN_FETCH_QUERY,
+        block: suspend CoroutineScope.(OnSuggestedDomains) -> T
+    ) {
+        whenever(purchasingFeatureConfig.isEnabledOrManuallyOverridden()).thenReturn(true)
+        test {
+            val (query, size) = queryToSize
+
+            val suggestions = (0 until size).map {
+                DomainSuggestionResponse().apply {
+                    domain_name = "$query-$it.com"
+                    is_free = it % 2 == 0
+                    cost = if (is_free) "Free" else "$$it.00"
+                }
+            }
+
+            val event = OnSuggestedDomains(query, suggestions)
+            whenever(fetchDomainsUseCase.fetchDomains(any(), any(), any(), any())).thenReturn(event)
+            block(event)
+        }
+    }
+
+    private val uiDomains get () = assertIs<List<New.DomainUiState>>(viewModel.uiState.value?.contentState?.items)
+
+    @Test
+    fun `verify all domain results from api visible`() = testWithSuccessResultNewUi { (query, results) ->
+        viewModel.start()
+
+        viewModel.onQueryChanged(query)
+        advanceUntilIdle()
+
+        assertThat(uiDomains).hasSameSizeAs(results)
+    }
+
+    @Test
+    fun `verify cost of free domain results from api is 'Free'`() = testWithSuccessResultNewUi { (query, results) ->
+        val apiFreeDomains = results.filter { it.is_free }
+        viewModel.start()
+
+        viewModel.onQueryChanged(query)
+        advanceUntilIdle()
+
+        assertThat(uiDomains).filteredOn { it.cost is Cost.Free }.hasSameSizeAs(apiFreeDomains)
+    }
+
+    @Test
+    fun `verify cost of paid domain results from api is 'Paid'`() = testWithSuccessResultNewUi { (query, results) ->
+        val apiPaidDomains = results.filter { !it.is_free }
+        viewModel.start()
+
+        viewModel.onQueryChanged(query)
+        advanceUntilIdle()
+
+        assertThat(uiDomains).filteredOn { it.cost is Cost.Paid }.hasSameSizeAs(apiPaidDomains)
+    }
+
+    @Test
+    fun `verify only 1st domain result from api is 'Recommended'`() = testWithSuccessResultNewUi { (query) ->
+        viewModel.start()
+
+        viewModel.onQueryChanged(query)
+        advanceUntilIdle()
+
+        assertThat(uiDomains.map { it.variant }).containsOnlyOnce(New.DomainUiState.Variant.Recommended)
+    }
+
+    @Test
+    fun `verify only 2nd domain result from api is 'BestAlternative'`() = testWithSuccessResultNewUi { (query) ->
+        viewModel.start()
+
+        viewModel.onQueryChanged(query)
+        advanceUntilIdle()
+
+        assertThat(uiDomains.map { it.variant }).containsOnlyOnce(New.DomainUiState.Variant.BestAlternative)
+    }
+
+    // endregion
 }
 
 /**
